Checkpoint the database at version 53
I've made some pretty significant changes lately converting character set/
collation to something more modern and setting the engine for all tables to
InnoDB so I think it makes sense to do this.

The proximate motivation was a bunch of issues with Foreign Keys when trying
to bootstrap the dockerized dev env. This way we skip all that and jump
straight to the current working schema.