Change of approach to some of the database stuff
The motivating change here is to write some tests that care (at least a little
about the state of the database). Right now we rely upon you importing M10,
ELD and running updateDefaultFormats. That's fine as far as it goes but it's
kinda temporal. What's legal in Standard in your tests will change depending on
when you set up your local db.

So here we introduce the idea of a test db that's separate from your dev db. We
build it from a known .sql file, and then apply any migrations that have been
written since it was dumped. It's possible that keeping this "up to date" with
tests will be a lot of hassle but for right now it gives us a moment in time
that we can code some tests against. We'll see how it goes.

The test db is created once only for each test suite run (because it takes a
couple seconds and we don't want to add that overhead to every single test) but
a form of isolation between tests is created by opening a transaction before
each test and then rolling it back. Tests that issue DDL (should be very few)
will break this a little but it should give pretty good test isolation while
remaining fast. (This work done by DatabaseTestListener.)

In the creation of this I ended up doing rather a lot:

  - A new way of issuing db queries through Gatherling\Data\DB. I think the
    interface here is nicer. It hides what PHP lib we're using from the caller.
    It only requires you to handle our exceptions, not PHP's. It supports
    transactions (at least as much as mariadb does) and fakes support for
    nested transactions, too, using SAVEPOINT. (It's possible that we should not
    SAVEPOINT but let's try it like this for now.) Converting existing db stuff
    to use this is a future plan.
  - A DatabaseException for "something went wrong with the database".
  - A FileNotFoundException, currently just for when we're not able to load
    mysqldump files.
  - Change how we do migrations. The old hardcoded way replaced by numbered
    pure SQL files in a migrations dir. Everything is essentially checkpointed
    at version 51 although git will remain the details should it ever matter.
  - A Setup class that knows how to get you from your current state to a working
    db and/or test db.
  - Configuration options for test db which lightly encourage you to use the
    same user for your dev and test db but keep them separate (I think that's
    ok?)
  - Adding a new simple migration mostly just to show how they work now.
  - Adding LogTestListener that buffers log output during test running and only
    outputs when a test fails.