Let calls to DB::value indicate if the value not being present is ok or not
If it's not ok (default) then throw a DatabaseException. If it is ok then return null when the query gets no result.

Note: there is no way to differentiate between "found and was NULL" and "not found" but that's probably better than the previous situation of not being able to tell if it was "found and is FALSE". If we need the former we'll have to complicate this later.

Included - change to checkPassword which uses the new facility and thus avoids using a variable that might be null.