Convert ratings.php to use a Page
Definitely some imperfections here and the handling of Pagination leaves a lot
to be desired but I think this is a good step towards having everything in the
Page/Component system which will let us do centralized routing/error handling
and gets the echoes/XSS vulnerabilities out of the system.

It's possible we want to move the db access out of the controllers but that is
left for another day.