Introduce static typechecking with PHPStan
I had a couple of type issues in new code so I thought it was time to
introduce typechecking.

I used the --generate-baseline option to exclude a whole bunch of stuff that's
grandfathered in, and fixed a bunch of egregious issues that were simple to fix.

Level 5 is as far as we can sensibly go for now because Level 6 requires
everything to be typehinted and it would take too long to even read all those
error messages.

Future work involves:

    (1) Bringing the old code that doesn't get removed up to level 5.
    (2) Bringing the whole codebase up to level 6 and beyond.