Don't bind everything in SQL query as a string, inspect type
The immediate reason for this is to be able to bind a LIMIT param without
having to complicate the interface for `select`.

We may want to make this explicit not automatic, or we may want to expand the
sniffing to other types, but for now this will allow me to pass a LIMIT param
as an int and not have it be quoted and mess up the resutling query.