Baseline an ungodly amount of phpstan issues, trying to get access to level 9
We have fatal type errors from assigning bad values out of $_POST to ?int which
are only detected by phpstan at level 9 with bleeding edge enabled for some
reason. So here we are getting to the point where we can use those to find
actual site-breaking bugs by ignoring a bunch of stuff that probably should be
fixed but isn't going to blow up the world.