Move to phpstan level 9 by baselining a lot of errors
Unfortunately bleeding edge + level 9 (the maximum) is the only place phpstan
will tell you about this:

class Test {
     public ?int $foo;
}

$bar = Test();
$bar->foo = $_POST['foo'];

which is a PHP Fatal Error with strict types on.

So we have baselined a whole bunch of stuff that should be looked at so that if
we ever introduce a sneaky PHP Fatal like the above phpstan will instantly tell
us about it when we run `composer static`.

We are not in any way close to level 9 and this may prove to be very annoying,
but at least we won't ship these very easy to ship fatals.

In the future we should stop accessing superglobal and go through some typesafe
intermediary like a Request object or something, but as there are more than 500
references to superglobals in the code that isn't happening instanter.