Band aid for a league result reporting issue
The behavior here is to take you back to print_submit_resultForm which isn't
even used for league reporting (in a league report the first player can only
report that they have won, to avoid a race condition with both players
reporting separately). So even if there were a match_id or something else to key
off this isn't taking you to the right place.

For sure we can do better here but for now we just hide a button that leads you
astray.

In the pre-type safety world this was not a PHP error. But probably very weird
and pointless things happened when you submitted the resulting form. So I'm
going to say this is a very minor win for the new way of doing things.