Don't crash if we try to remove an event entry that doesn't exist
We saw this on prod today. One way it can happen is if you have a stale page
with the "Unreg" link open and click after you already unregistered elsewhere.
It doesn't seem reasonable to explode in that situation even though it does no
harm so this treats it as a mildly notable log happening instead.