We must save the deck if we create or update it
Previously we were checking the return value of Deck->save which is always true.
In my desire to remove that I went too far and removed the actual calls to save
which we still need. We just no longer check the return value as it will always
be true. I'll convert this function to void at some point.