In order to make some better tests, convert a lot of the old db code
I have this nice DatabaseCase test parent class but using it is a pain because
as soon as you call some code that uses the old db style it doesn't know about
the transaction that isolates each test from each other test and things get
screwy.

So in order to add a couple of really very simple littles tests go kind of
crazy and rip out half of the old db calls in favor of the new style.

In the course of doing that discover that we need insert, insertMany and
update functions for one id, a list of ids, and a number of rows affected when
issuing db queries.

While dealing with that discover commonalities between Request and DB::value
and DB::values that made me want to factor out all the shared parts into what
is now Marshaller and the marshal helper.

Fix 100 or more type errors. Update the baseline. Onwards.