Rather scrappy func to interpolate SQL queries
When a query is failing or taking too long you really want to be able to
paste it into the mariadb commandline client and do an EXPLAIN and just run it.
Prepared statements while wonderful for security are not so good for that, so
here's something I mostly cribbed from Stack Overflow that lets you get the
query with the params interpolated for debugging.