Instead of making 60,000 queries to insert 15,000 legal cards, make three
The old system was exceeding Cloudflare's timeout when updating the Penny
Dreadul legal list from 0 to 15,000 legal cards. This version is a smidge
more complex but takes less than 1 second.

This probably implies a refactor of adding/removing ban list and restrited list,
too. But as in practicality it's only the legal list update that times out I've
left the rest on the backlog.