Mark as NOT NULL every column in the db that has no NULLs
It's possible this will break something where we allow NULLs initially but then
fill in the details shortly after, but I doubt it.

The idea here is just to bring the db schema in line with how it is actually
used which defends us against future bugs and data corruption.

There are a lot more things that could be improved but small moves in the right
direction help.