The title of the box (uppertitle) is the page title, so make it an h1
Except in the case where there is more than one box in a page. In that case
make the first uppertitle the h1, and the rest h2.

In theory this could result in an h2 "below" an h2 but in practice only the
homepage and decksearch have more than one h2 so it either doesn't happen or
happens very little.

We can probably do better here but at least all pages have one and only one h1
now.