Try and bring some consistency to our headings
h1 entitltes the page. There is always exactly one. It's a box .uppertitle.

h2 starts a subsection, of which there may be 0, 1 or many in a box/page.

h3 is a subsection of the subsection created by an h2.

We don't use h4, h5 or h6 yet but it would possibly make sense to make the
all caps grid subsection headings found on Player CP and for "ROUND" in Host CP
into h4s?

The purpose here is (at least) twofold - have the markup make semantic sense,
and have the site feel consistent and usable.