Move "Create New Event" out of "Filter Events" on Event List, and restyle
This button doesn't belong in "Filter Events". It doesn't have anything to do
with filtering events. It is a standalone option for a host to create a new
event. So it should stand alone.

For this button to stand alone it needs to be more prominent than current
.inputbutton styling. So take this opportunity to (a) restyle our buttons, and
(b) introduce the concept of primary button (the main button on the whole page,
there should never be more than one) versus just a "normal" button.

Then judiciously upgrade some existing buttons to primary. This is not a
complete survey. The normal new button styling is fine even for a primary
button so it's not important to hit them all exhaustively right this minute.
But in general from now on if the focus of a page is a form its main button
(but only one) should use btn-primary style.

As a result of this we have added a new variation on our link color that is used
for primary buttons. It's not as dark as "flame" but it's more prominent than
"Persian Orange". For stuff that's more important than one-of-many links but
not yet hovered.