Upgrade to PHP 8.2, switch to using "Safe" versions of built in funcs
It's annoying to have to check that preg_replace did not return null and that
ob_start did not return false. In these cases there's a catastrophic failure
not a normal run of events and we'd much prefer an exception be thrown and our
code be kept concise for normal operations.

Someone solved this for PHP7 with a library called "safe":

  - https://github.com/thecodingmachine/safe

The library rewrites builtin PHP funcs to throw exceptions instead of return
false or null or other funky behavior you have to check for in the edge cases.

Unfortunately they abandoned the library before PHP8 came out.

Luckily someone forked it and updated it for PHP8:

  - https://github.com/shish/safe

Unfortunately they only support PHP8.2+. But I have already written a bunch of
pointless code to workaround this problem, and there are tons of exceptions in
our static checker baselines to accommodate all this silliness.

So I upgraded us to PHP8.2 and updated all the code and removed some of the
pointless checking we were doing.