Remove FileNotFoundException
We were using this when we were explicitly checking file_get_contents. Now we
call the "safe" version that throws when it can't read the file. We lose a bit
of fidelity in the error type but our code is simpler. Now we aren't detecting
the error condition and throwing we don't need this.