Don't add a match every time Update Match Listing is clicked
Only add it if they filled out the dropdowns for a new match, otherwise they
might be deleting matches or reporting results, etc. and not trying to add a
new match.

This would've been caught by foreign keys only there's a player with the name
empty string gdi. Might have to re-org the db so an FK that stops you having a
match in a tournament you're not in is easier to declare.