Prep for a move to mobile-friendly responsive design
Right now if you remove the min-width the structural elements all move to a
fluid design that will work on mobile. But many of the individual pages then
break the right margin or become unusably squished. Checkpointing my work here
because it's a relevant milestone but the min-width means it's invisible until
we get the individual components behaving well enough to turn it on properly.