Don't tell someone they have the bye if they had the bye *last* round
You can get in a state where you've reported round N+1 and in round N you got
the bye and this logic would say "you have the bye in round N+1" which is not
true and is confusing.

This whole thing probably needs a revamp and the logic should be pushed into
the model or at least somewhere shared but for now let's not lie to people.