When we get a number that is a string in dictIntOrString, make it an int
This is better, I think, than having everything come out of dictIntOrString as
a string and rely on implicit conversion. Although '99.9' would come through as
a string, and 99.9 would raise. So it's weird. Possibly this should not exist.
It's only used for setting season rules right now and this is the behavior we
want there so I guess we will enshrine this behavior at least for now.