Error if admin tries to add a match to round 0
Previously the default value for the round dropdown was empty string and this
code would error out if you tried to submit the form without changing the
default.

Since making the default of RoundDropMenu zero instead of empty string (because
it should always be a valid int) this behavior changed to the round addition
being silently dropped. This is, I think, a smidge worse. So I've reinstated
erroring out until such time as we add real validation and nice error messages.