Treat empty string in optionalInt or optionalFloat as null
I did not apply this at the Marshaller level. This wouldn't apply to a db query
(where no row would be returned) or similar.

But in a querystring `series_id=1&season=' is better served by setting
get()->optionalInt('season') to null rather than throwing an exception.

It will certainly make returning eventreport.php to working order easier.

We *could* instead have the select set a magic constant for season=All of 'All'
or '99' or '-1' but that seems worse than this. (0 is already taken as it is
possible for a series to have a season 0 for some reason.)