Differentiate between 0 and null in NumDropMenu
The season dropdown on eventreport.php has both "All" (empty string, converted
to null by Request->optionalInt) and "0" (converted to 0 by
Request->optionalInt) both of which are fuzzily equal to 0 but only one of which
is strictly equal to 0. So take this opportunity to tighten up the types here
and make that work how it should ("All" is preselected when no season specified,
but 0 is preselected if season 0 is specified).