Split event display into Pending, Active, Upcoming and Past Events
Events that need to be started are most important, event that are running are
next most important, events in the future are next, and old events are least;
now the hierarchy of the page reflects that.

Remove a bunch of not-super-useful columns from event display. This'll work
better on mobile and "k value" and "host/cohost" were not super useful columns.
"Finalized" is now implied by location on the page and can go too.