Remove php-icons in favor of font awesome
I actually tried to go the other way, and remove font awesome in favor of
php-icons and it was a giant pain and nothing looked right. Replacing the three
icons with font awesome options was the work of a few seconds. So I guess that
tells me which way is better. This also has no "remember to scan for the icons
on prod" step, too.