Update font awesome in order to get the new(ish) Discord logo
This probably also loads a heck of a lot less stuff than the old way making
initial page load faster.

This is hidden behind a login but it's just the generic free font awesome so
hopefully that isn't a problem and we don't need to configure anything. Or just
swap it out if you do.