Use DateTimeImmutable rather than ints to represent dates in more places
Add a `datetime` helper that lets us convert easily between the string we get
from the db code in the DTO into a DateTimeImmutable. (Unfortunately it's not
simple to have PDO's FETCH_CLASS do this for us.)

In general move towards a system that passes around DateTimeImmutable rather
than int for dates.