Don't allow entries with matches to be removed, prompt to drop instead
This relates to issue 908. We find orphaned decks and entries with NULL deck
which then causes an exception to be thrown when showing them. Here we make it
so that if you submit a stale unreg link you get prompted to drop instead of
leaving the db in a bad state.

If an admin or the API attempts something similar they will just get a 500 and
no slick handling but I think that's probably rare enough that it's ok, and
admins have more recourse if they get a 500, too.

Not sure this is the complete fix but we can check in a short while if the
orphaning has stopped happening.