Now the lib.php doesn't really do anything move it into bootstrap and delete
This used to have a ton of shared code which was its main purpose as well as a
bit of setup. Now it's just that smidge of setup left and all the shared code
is in components or helpers. So we can combine with bootstrap.php and let
lib.php go.