Don't show the login box to already-signed-in folks
This was relying on Page->player that I removed when making the display name
shorter for the logout button in the menu.

It's maybe slightly odd to rely on this truncate version of the username as
the "is signed in or not" indicator, but I renamed it so it makes at least a
little more sense. No point in having a whole extra property just for this I
don't think.