Remove bogus round param from various places
This was effectively being used as a boolean - 0 means don't bother checking the
database just return an empty list, and any other value means "go look in the
db". I think it's very misleading to say getOpponents(round=1) and see your
opponents for every round. So even though it means reading the db unnecessarily
before the tournament starts that has to be fast enough anyway for the scenarios
where the tournament has started so I'm removing the misleading arg and
simplifying the code.