Add a convenient flag to update test db
And instructions on how to use it to checkpoint test-db.sql.

It's not clear to me if it was this or the move to using the commandline client
instead of splitting the dump and executing one-by-one from PHP but one of the
two things (or a combination) reduced the time of `composer test` from 12-15s
to 8-9s on my laptop.