Restore database dump via the commandline client rather than from PHP
This way if the dump has any DELIMITER in it (as is output for TRIGGERs) the
restore will still work. Needed this because of 86.sql migration.

It's not clear if it was this or the checkpointing of the database but between
them they made the time of `composer test` go from 12-15s to 8-9s on my laptop.